Add import component
