+++ .gitignore
@@ -169,3 +169,4 @@ cache_all_tariffs.pkl
 .aws/config
 .aws/credentials
 node_modules
+frontend/package-lock.json
