+++ frontend/src/api/certificateAPI.js
@@ -16,3 +16,17 @@ export const cancelCertificateAPI = (certificateData) =>
 export const getCertificateDetailsAPI = (certificateId) => {
   return baseAPI.get(`/certificate/${certificateId}`);
 };
+
+export const downloadCertificateImportTemplateAPI = () => {
+  return baseAPI.get("/certificate/certificate_import_template", {
+    responseType: "blob",
+  });
+};
+
+export const importCertificatesAPI = (formData) => {
+  return baseAPI.post("/certificate/import", formData, {
+    headers: {
+      "Content-Type": "multipart/form-data",
+    },
+  });
+};
