--- frontend/src/components/device/Device.module.css
@@ -1,8 +1,0 @@
-.card-body {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  height: 168px;
-  gap: 24px;
-  width: 100%;
-}
